Fix some signed/unsigned variable comparison warnings by changing the "bitmode" from a long to a DWORD.  This should have no effect on the code that the compiler generates.
git-svn-id: https://svn.code.sf.net/p/fmslogo/source@1747 9d0050d7-00b0-40f9-871c-8e02cbd5794f

+++ trunk/logo/src/globals.h
@@ -187,7 +187,7 @@ struct Turtle
    MATRIX   Matrix;
    VECTOR   Position;
    FLONUM   Heading;
-   long     Bitmap;
+   DWORD    Bitmap;
    bool     IsShown;
    bool     IsPenUp;
    };
